# src/main/webapp/openapi.yaml
openapi: 3.0.3
info:
  title: Soccorso Web API
  version: 1.0.0
  description: >
    Questa API permette di gestire un sistema di soccorso in situazioni di emergenza.
    Tutti possono inserire richieste di aiuto.
    Gli admin e gli operatori possono autenticarsi, validarle, e monitorare
    lo stato delle missioni create per rispondere a queste richieste. Gli admin possono anche annullare richieste, creare missioni e chiuderle valutando il successo
    dell'intervento. Tutto è gestito in memoria per fare il test dell'API.
    È disponibile anche un client web semplice per testare le funzionalità principali.
servers:
  - url: http://localhost:8080/SoccorsoWebServices/resources
    description: URL API
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: UUID
      description: Token di autenticazione UUID da includere nell’header Authorization
  schemas:
    Request:
      type: object
      description: Rappresenta una richiesta di soccorso.
      properties:
        id:
          type: integer
          format: int64
          description: Identificatore univoco della richiesta
          example: 1234
        status:
          type: string
          description: Stato attuale della richiesta
          example: ACTIVE, IN_PROGRESS, CLOSED, IGNORED
          default: ACTIVE
        successLevel:
          type: integer
          nullable: true
          minimum: 0
          maximum: 10
          description: Livello di successo associato alla gestione della richiesta
          example: 7
          default: null
        type:
          type: string
          description: Tipo di richiesta (MEDICA, INCENDIO, TERREMOTO, ...)
          example: "INCENDIO"
        location:
          type: string
          description: Localizzazione dell'evento
          example: "Via Verdi 12, Roma"
        requestTime:
          type: string
          format: date-time
          description: Data e ora in cui la richiesta è stata registrata
          example: "2025-07-18 20:25:48"
        validated:
          type: boolean
          description: Indica se la richiesta è stata validata da un operatore
          example: false
          default: false
      required:
        - type
        - location
...
paths:
...
  /requests:
    post:
      tags:
        - HelpRequests
      summary: Crea una nuova richiesta di soccorso
      description: |
        Crea una nuova richiesta di soccorso con stato iniziale `ACTIVE`.
        Non richiede autenticazione.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Request'
            example:
              Type: "INCENDIO"
              Location: "Via Verdi 12, Roma"
      responses:
        '201':
          description: Richiesta creata con successo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Request'
              example:
                id: 1234
                status: ACTIVE
                successLevel: 7
                type: MEDICA
                location: Roma
                requestTime: '2025-07-18 20:25:48'
                validated: false
              
        '400':
          description: Dati non validi
          
    get:
      tags:
        - HelpRequests
      summary: Elenco richieste filtrate per stato passato tramite parametro query (paginato)
      description: |
        Restituisce l'elenco delle richieste di soccorso con stato specificato.
        Richiede autenticazione. I risultati sono paginati. Non da 404 perché verrà semplicemente fornito un array vuoto.
      security:
        - BearerAuth: []
      parameters:
        - name: status
          in: query
          required: true
          schema:
            type: string
            example: ACTIVE, IN_PROGRESS, CLOSED
          description: Stato della richiesta da filtrare (es. ACTIVE, CLOSED, ecc.)
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 1
          description: Numero della pagina (da 1 in poi)
        - name: size
          in: query
          required: false
          schema:
            type: integer
            default: 15
          description: Numero massimo di risultati per pagina
      responses:
        '200':
          description: Elenco richieste restituito con successo
          headers:
            X-Total-Count:
              description: Numero totale di richieste che soddisfano il filtro
              schema:
                type: integer
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Request'
                example:
                    - id: 1
                      status: CLOSED
                      successLevel: 4
                      type: INCENDIO
                      location: Pescocostanzo
                      requestTime: "2024-06-30 18:25:48"
                      validated: true
                    - id: 2
                      status: IN_PROGRESS
                      successLevel: null
                      type: ALLUVIONE
                      location: Sulmona
                      requestTime: "2025-06-30 18:25:48"
                      validated: true
        '400':
          description: Parametri non validi
        '401':
          description: Token non fornito o non valido
  /requests/{id}:
    get:
      tags:
        - HelpRequests
      summary: Ottieni una richiesta per ID
      description: Restituisce i dettagli della richiesta specificata. Richiede autenticazione.
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Richiesta trovata
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Request'
              example:
                id: 1
                Status: IGNORED
                SuccessLevel: null
                Type: POLIZIA
                Location: Milano
                RequestTime: '2024-03-15 20:25:48'
                Validated: false
        '404':
            description: Richiesta non trovata
        '401':
          description: Token assente o non valido
  /requests/{id}/validate:
    get:
      tags:
        - HelpRequests
      summary: Valida una richiesta di soccorso
      description: |
        Convalida una richiesta (se non già validata).
        Richiede autenticazione.
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Richiesta validata con successo
        '404':
          description: Richiesta non trovata o già validata
        '401':
          description: Token non valido
  /requests/{id}/remove:
    delete:
      tags:
        - HelpRequests
      summary: Elimina una richiesta di soccorso
      description: |
        Rimuove la richiesta se esiste. Richiede ruolo admin.
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Richiesta eliminata con successo
        '404':
          description: Richiesta non trovata
        '403':
          description: Operazione consentita solo ad admin
        '401':
          description: Token non valido
  /requests/{id}/ignore:
    put:
      tags:
          - HelpRequests
      summary: Ignora una richiesta di soccorso
      description: | 
       Se ad esempio una richiesta di soccorso viene giudicata falsa può non essere validata e ignorata. Richiede ruolo *admin*.
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Richiesta eliminata con successo
        '404':
          description: Richiesta non trovata
        '403':
          description: Operazione consentita solo ad admin
        '401':
          description: Token non valido
  /requests/notPositive:
    get:
      tags:
        - HelpRequests
      summary: Ottieni richieste con successo ≤ 5
      description: |
        Restituisce tutte le richieste con successo basso (≤5).
        Richiede autenticazione.
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Richieste trovate
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Request'
                example:
                    - id: 4
                      status: CLOSED
                      successLevel: 4
                      type: INCENDIO
                      location: Pescocostanzo
                      requestTime: "2024-12-30 18:25:48"
                      validated: true
                    - id: 5
                      status: CLOSED
                      successLevel: 2
                      type: ALLUVIONE
                      location: Tocco da Casauria
                      requestTime: "2025-12-13 18:25:48"
                      validated: true
        '401':
          description: Token assente o non valido
        '404':
          description: Nessuna richiesta con successo basso trovata
...